Unit tests compilation and Rust-fmt. Matchers tests moved from network.rs to network_matchers.rs
