+++ tests/unit/optimizer.rs
@@ -3,9 +3,9 @@ mod optimization_tests_pattern_group {
     #[cfg(test)]
     mod optimization_tests_pattern_group {
         use super::*;
-        use crate::filters::network::CompiledRegex;
         use crate::filters::network::NetworkMatchable;
         use crate::lists;
+        use crate::regex_manager::CompiledRegex;
         use crate::regex_manager::RegexManager;
         use crate::request::Request;
         use regex::bytes::RegexSetBuilder as BytesRegexSetBuilder;
@@ -331,9 +331,9 @@ mod optimization_tests_pattern_group {
     }
     */
     use super::super::*;
-    use crate::filters::network::CompiledRegex;
     use crate::filters::network::NetworkMatchable;
     use crate::lists;
+    use crate::regex_manager::CompiledRegex;
     use crate::regex_manager::RegexManager;
     use crate::request::Request;
     use regex::bytes::RegexSetBuilder as BytesRegexSetBuilder;
